# src/api/docs/openapi.yaml
openapi: 3.1.0
info:
  title: GUI-LOP API
  description: |
    ## Generative UI & Human-in-the-Loop Orchestration Platform API

    The GUI-LOP API provides comprehensive endpoints for authentication, workflow management,
    real-time communication, and system monitoring. This API supports:

    - JWT-based authentication with refresh tokens
    - Workflow creation, execution, and management
    - Real-time WebSocket communication
    - Rate limiting and security features
    - Comprehensive error handling
    - API versioning and backward compatibility

    ### Authentication
    The API uses JWT (JSON Web Tokens) for authentication. Include the access token
    in the Authorization header:

    ```
    Authorization: Bearer <access_token>
    ```

    ### Rate Limiting
    API endpoints are rate-limited to prevent abuse. Rate limit information is
    included in response headers:

    - `X-RateLimit-Limit`: Maximum requests per window
    - `X-RateLimit-Remaining`: Remaining requests in current window
    - `X-RateLimit-Reset`: Time when rate limit resets (Unix timestamp)

    ### Error Responses
    All error responses follow a consistent format:

    ```json
    {
      "success": false,
      "message": "Error description",
      "code": "ERROR_CODE",
      "details": {...},
      "timestamp": "2024-01-01T00:00:00.000Z"
    }
    ```

    ### Pagination
    List endpoints support pagination using:
    - `page`: Page number (default: 1)
...

...
    ```json
    {
...
      }
    }
...
servers:
...
security:
  - BearerAuth: []
  - ApiKeyAuth: []
...
paths:
...
  /auth/register:
    post:
      tags: [Authentication]
      summary: Register a new user
      description: |
        Create a new user account with email and password. The password must meet
        security requirements: at least 8 characters with uppercase, lowercase,
        number, and special character.
      operationId: registerUser
      security: [] # No authentication required
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRegistrationRequest'
            examples:
              standardUser:
                summary: Standard user registration
                value:
                  email: "user@example.com"
                  password: "SecurePass123!"
                  confirmPassword: "SecurePass123!"
                  firstName: "John"
                  lastName: "Doe"
                  role: "user"
              adminUser:
                summary: Admin user registration (requires special permissions)
                value:
                  email: "admin@example.com"
                  password: "AdminPass123!"
                  confirmPassword: "AdminPass123!"
                  firstName: "Admin"
                  lastName: "User"
                  role: "admin"
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
              examples:
                successfulRegistration:
                  summary: Successful registration response
                  value:
                    success: true
                    message: "User registered successfully"
                    data:
                      user:
                        id: "user_1234567890_abcdef"
                        email: "user@example.com"
                        firstName: "John"
                        lastName: "Doe"
                        role: "user"
                        isActive: true
                        createdAt: "2024-01-01T00:00:00.000Z"
                      tokens:
                        accessToken: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                        refreshToken: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                        expiresIn: 3600
                    metadata:
                      timestamp: "2024-01-01T00:00:00.000Z"
        '400':
          description: Bad request - validation failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                validationError:
                  summary: Validation error example
                  value:
                    success: false
                    message: "Validation failed"
                    code: "VALIDATION_ERROR"
                    details:
                      - field: "email"
                        message: "Valid email is required"
                      - field: "password"
                        message: "Password must be at least 8 characters"
                    timestamp: "2024-01-01T00:00:00.000Z"
        '409':
          description: Conflict - user already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Too many requests - rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/login:
    post:
      tags: [Authentication]
      summary: Authenticate user
      description: |
        Authenticate a user with email and password. Returns JWT access and refresh tokens.
        Failed attempts are tracked and accounts may be temporarily locked after
        multiple failed attempts.
      operationId: loginUser
      security: [] # No authentication required
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
            examples:
              standardLogin:
                summary: Standard user login
                value:
                  email: "user@example.com"
                  password: "SecurePass123!"
                  rememberMe: false
              rememberedLogin:
                summary: Login with remember me
                value:
                  email: "user@example.com"
                  password: "SecurePass123!"
                  rememberMe: true
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Bad request - validation failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Authentication failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '423':
          description: Account locked
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Too many requests
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/refresh:
    post:
      tags: [Authentication]
      summary: Refresh access token
      description: |
        Refresh an expired access token using a valid refresh token. Refresh tokens
        are single-use and will be invalidated after use.
      operationId: refreshToken
      security: [] # No authentication required
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenRequest'
      responses:
        '200':
          description: Token refreshed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RefreshTokenResponse'
        '400':
          description: Bad request - validation failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Invalid or expired refresh token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Too many requests
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/logout:
    post:
      tags: [Authentication]
      summary: Logout user
      description: |
        Logout the current user and revoke tokens. Accepts optional refresh token
        to revoke specific session. All user tokens can be revoked for complete logout.
      operationId: logoutUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LogoutRequest'
      responses:
        '200':
          description: Logout successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LogoutResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/me:
    get:
      tags: [Authentication]
      summary: Get current user profile
      description: Retrieve the profile information of the currently authenticated user.
      operationId: getCurrentUser
      responses:
        '200':
          description: User profile retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfileResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/change-password:
    post:
      tags: [Authentication]
      summary: Change user password
      description: |
        Change the current user's password. Requires current password for verification
        and will revoke all existing tokens, forcing re-login.
      operationId: changePassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangePasswordRequest'
      responses:
        '200':
          description: Password changed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChangePasswordResponse'
        '400':
          description: Bad request - validation failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Current password incorrect
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # Workflow Endpoints
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT access token obtained from /auth/login or /auth/refresh endpoints.
        Include in Authorization header: "Bearer <token>"

    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for service-to-service authentication
...
  schemas:
...
    UserRegistrationRequest:
      type: object
      required: [email, password, confirmPassword, firstName, lastName]
      properties:
        email:
          type: string
          format: email
          description: User email address
          example: "user@example.com"
        password:
          type: string
          minLength: 8
          description: |
            Password must be at least 8 characters with uppercase, lowercase,
            number, and special character
          example: "SecurePass123!"
        confirmPassword:
          type: string
          description: Password confirmation
          example: "SecurePass123!"
        firstName:
          type: string
          minLength: 1
          maxLength: 50
          description: First name
          example: "John"
        lastName:
          type: string
          minLength: 1
          maxLength: 50
          description: Last name
          example: "Doe"
        role:
          type: string
          enum: [user, admin]
          default: user
          description: User role
          example: "user"
      additionalProperties: false
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
          description: User email address
          example: "user@example.com"
        password:
          type: string
          description: User password
          example: "SecurePass123!"
        rememberMe:
          type: boolean
          default: false
          description: Keep user logged in longer
          example: false
      additionalProperties: false
    RefreshTokenRequest:
      type: object
      required: [refreshToken]
      properties:
        refreshToken:
          type: string
          description: Valid refresh token
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
      additionalProperties: false
    LogoutRequest:
      type: object
      properties:
        refreshToken:
          type: string
          description: Optional refresh token to revoke specific session
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        revokeAll:
          type: boolean
          default: false
          description: Revoke all user tokens
          example: false
      additionalProperties: false
    ChangePasswordRequest:
      type: object
      required: [currentPassword, newPassword, confirmPassword]
      properties:
        currentPassword:
          type: string
          description: Current password for verification
          example: "OldPass123!"
        newPassword:
          type: string
          minLength: 8
          description: |
            New password must be at least 8 characters with uppercase, lowercase,
            number, and special character
          example: "NewPass123!"
        confirmPassword:
          type: string
          description: New password confirmation
          example: "NewPass123!"
      additionalProperties: false
...
    AuthResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: "Login successful"
        data:
          type: object
          properties:
            user:
              $ref: '#/components/schemas/User'
            tokens:
              $ref: '#/components/schemas/Tokens'
          required: [user, tokens]
        metadata:
          $ref: '#/components/schemas/ResponseMetadata'
      required: [success, message, data, metadata]
    RefreshTokenResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: "Token refreshed successfully"
        data:
          type: object
          properties:
            tokens:
              $ref: '#/components/schemas/Tokens'
          required: [tokens]
        metadata:
          $ref: '#/components/schemas/ResponseMetadata'
      required: [success, message, data, metadata]
    LogoutResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: "Logout successful"
        data:
          type: object
          properties:
            revokedAt:
              type: string
              format: date-time
              example: "2024-01-01T00:00:00.000Z"
          required: [revokedAt]
        metadata:
          $ref: '#/components/schemas/ResponseMetadata'
      required: [success, message, data, metadata]
    ChangePasswordResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: "Password changed successfully"
        data:
          type: object
          properties:
            changedAt:
              type: string
              format: date-time
              example: "2024-01-01T00:00:00.000Z"
          required: [changedAt]
        metadata:
          $ref: '#/components/schemas/ResponseMetadata'
      required: [success, message, data, metadata]
    UserProfileResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: "User profile retrieved successfully"
        data:
          type: object
          properties:
            user:
              $ref: '#/components/schemas/User'
          required: [user]
        metadata:
          $ref: '#/components/schemas/ResponseMetadata'
      required: [success, message, data, metadata]
...
    Tokens:
      type: object
      properties:
        accessToken:
          type: string
          description: JWT access token
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        refreshToken:
          type: string
          description: JWT refresh token
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        expiresIn:
          type: integer
          description: Access token expiration time in seconds
          example: 3600
        tokenType:
          type: string
          enum: [Bearer]
          example: "Bearer"
      required: [accessToken, expiresIn, tokenType]
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          description: Human-readable error message
          example: "Request validation failed"
        code:
          type: string
          description: Machine-readable error code
          example: "VALIDATION_ERROR"
        details:
          oneOf:
            - type: array
              items:
                type: object
                properties:
                  field:
                    type: string
                    description: Field with error
                    example: "email"
                  message:
                    type: string
                    description: Error message
                    example: "Valid email is required"
                  code:
                    type: string
                    description: Error code for field
                    example: "INVALID_EMAIL"
                required: [field, message]
            - type: object
              additionalProperties: true
          description: Additional error details
        timestamp:
          type: string
          format: date-time
          description: Error timestamp
          example: "2024-01-01T00:00:00.000Z"
        requestId:
          type: string
          format: uuid
          description: Request identifier
          example: "123e4567-e89b-12d3-a456-426614174000"
        path:
          type: string
          description: Request path
          example: "/api/v1/auth/login"
        retryAfter:
          type: integer
          description: Seconds to wait before retrying
          example: 60
      required: [success, message, code, timestamp]

  # Reusable Request Bodies
...